requirements.txt
pinecone-client>=3.0.0
openai>=1.0.0
xai-sdk>=0.1.0
notion-client>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
tiktoken>=0.5.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
